requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dominate==2.4.0     
filelock==3.0.12
gensim==3.8.0       
h5py==2.8.0
idna==2.10
ipdb==0.12.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
numpy==1.18.1
packaging==20.7
pandas==1.0.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-warmup==0.0.4   
pytz==2020.4
regex==2020.11.13
requests==2.25.0
sacremoses==0.0.43
scipy==1.4.1
sentencepiece==0.1.91   
six==1.15.0
smart-open==4.0.1     
tensorboardX==2.0
tokenizers==0.8.1rc2
tqdm==4.54.1
traitlets==5.0.5
transformers==3.1.0
typeguard==2.7.0
urllib3==1.26.2
wcwidth==0.2.5
